+++ DTO/heSoDTO.cs
@@ -6,11 +6,11 @@ namespace DTO
     {
         public int MaNguyenLieu { get; set; }
         public int MaDonVi { get; set; }
-        public float HeSo { get; set; }
+        public decimal HeSo { get; set; }
 
         public heSoDTO() { }
 
-        public heSoDTO(int maNguyenLieu, int maDonVi, float heSo)
+        public heSoDTO(int maNguyenLieu, int maDonVi, decimal heSo)
         {
             MaNguyenLieu = maNguyenLieu;
             MaDonVi = maDonVi;
